Add Dockerfile
This image runs org.brann.clock.test.Testbed with whatever arguments the
user wants. They will need to mount a volume containing their input file.

This is a bit odd because the class that has our main is in the test JAR
so we have to put both in there. This was easier then trying to make maven
put all the files into one JAR.

Change-Id: I752a66b06a3fba7b59f14d6cd89b1959ce27fadc